fix msvc quirks
is new compiler used? not sure, but new install seems to have prompted stricter code

+++ src/NetAcceptorSteamDedicated.cpp
@@ -52,7 +52,7 @@ AcceptorSteam::AcceptorSteam() {
         LOG_INFO(LOGGER, "Logged into steam as {}", SteamFriends()->GetPersonaName());
     }
 
-    LOG_INFO(LOGGER, "Authentication status: {}", STEAM_NETWORKING_SOCKETS->InitAuthentication());
+    LOG_INFO(LOGGER, "Authentication status: {}", (int) STEAM_NETWORKING_SOCKETS->InitAuthentication());
     
     auto timeout = (float)duration_cast<milliseconds>(Valhalla()->Settings().playerTimeout).count();
     int32 offline = 1;
